CPR-615 Handle needs attention scenarios
